+++ CHANGELOG.md
@@ -4,6 +4,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.35.0] - 2018-10-15
+### Added
+- Added `encoding/protobuf/reflection` for exposing server reflection related
+  information through codegeneration. For docs related to server reflection read
+  https://github.com/grpc/grpc/blob/master/doc/server-reflection.md.
+- `protoc-gen-yarpc-go` now generates `yarpcfx` fx groups containing
+  information required for building server reflection API's.
+
+### Fixed
+- Using a `http.Outbound` previously leaked implementation details that it was using a
+    `*http.Client` underneath, when attempting to cast a `http.Outbound` into a `http.RoundTripper`
+
 ## [1.34.0] - 2018-10-03
 ### Added
 - Adds `thrift.Named` option for appropriately labelling procedures inherited
@@ -1022,6 +1034,7 @@ This release requires regeneration of ThriftRW code.
 
 - Initial release.
 
+[1.35.0]: https://github.com/yarpc/yarpc-go/compare/v1.34.0...v1.35.0
 [1.34.0]: https://github.com/yarpc/yarpc-go/compare/v1.33.0...v1.34.0
 [1.33.0]: https://github.com/yarpc/yarpc-go/compare/v1.32.4...v1.33.0
 [1.32.4]: https://github.com/yarpc/yarpc-go/compare/v1.32.3...v1.32.4